backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
cloudinary==1.36.0
pymongo==4.5.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
Pillow==10.0.0
requests==2.31.0